Rename Subject Header Component
We need more of these, and I don't think it quite fits under the subject
tree.
