+++ packages/frontend/app/components/reports/subject.hbs
@@ -9,7 +9,7 @@
         {{t "general.backToReports"}}
       </LinkTo>
     </div>
-    <Reports::Subject::Header
+    <Reports::SubjectHeader
       @report={{@report}}
       @school={{this.school}}
       @year={{@year}}
